+++ src/main/java/org/spongepowered/common/registry/AdditionalRegistration.java
@@ -30,16 +30,7 @@ import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
 @Retention(RetentionPolicy.RUNTIME)
-@Target(ElementType.TYPE)
-public @interface Registration {
-
-    Phase value();
-
-    enum Phase {
-        PRE_INIT,
-        INIT,
-        POST_INIT,
-        ;
-    }
+@Target(ElementType.METHOD)
+public @interface AdditionalRegistration {
 
 }
