--- src/main/java/org/spongepowered/common/registry/BaseRegistry.java
@@ -1,7 +1,0 @@
-package org.spongepowered.common.registry;
-
-public interface BaseRegistry {
-
-    void registerValues();
-
-}
