--- src/main/java/org/spongepowered/common/registry/CatalogRegistry.java
@@ -1,14 +1,0 @@
-package org.spongepowered.common.registry;
-
-import org.spongepowered.api.CatalogType;
-
-import java.util.Collection;
-import java.util.Optional;
-
-public interface CatalogRegistry<T extends CatalogType> extends BaseRegistry {
-
-    Optional<T> getById(String id);
-
-    Collection<T> getAll();
-
-}
