+++ src/main/java/org/spongepowered/common/registry/CatalogRegistryModule.java
@@ -24,22 +24,15 @@
  */
 package org.spongepowered.common.registry;
 
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
+import org.spongepowered.api.CatalogType;
 
-@Retention(RetentionPolicy.RUNTIME)
-@Target(ElementType.TYPE)
-public @interface Registration {
+import java.util.Collection;
+import java.util.Optional;
 
-    Phase value();
+public interface CatalogRegistryModule<T extends CatalogType> extends RegistryModule {
 
-    enum Phase {
-        PRE_INIT,
-        INIT,
-        POST_INIT,
-        ;
-    }
+    Optional<T> getById(String id);
+
+    Collection<T> getAll();
 
 }
