+++ src/main/java/org/spongepowered/common/registry/RegisterCatalog.java
@@ -24,6 +24,8 @@
  */
 package org.spongepowered.common.registry;
 
+import org.spongepowered.api.CatalogType;
+
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
@@ -31,15 +33,8 @@ import java.lang.annotation.Target;
 
 @Retention(RetentionPolicy.RUNTIME)
 @Target(ElementType.TYPE)
-public @interface Registration {
+public @interface RegisterCatalog {
 
-    Phase value();
-
-    enum Phase {
-        PRE_INIT,
-        INIT,
-        POST_INIT,
-        ;
-    }
+    Class<? extends CatalogType> value();
 
 }
