+++ src/main/java/org/spongepowered/common/registry/type/GoldenAppleRegistryModule.java
@@ -1,0 +1,66 @@
+/*
+ * This file is part of Sponge, licensed under the MIT License (MIT).
+ *
+ * Copyright (c) SpongePowered <https://www.spongepowered.org>
+ * Copyright (c) contributors
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in
+ * all copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+ * THE SOFTWARE.
+ */
+package org.spongepowered.common.registry.type;
+
+import static com.google.common.base.Preconditions.checkNotNull;
+
+import com.google.common.collect.ImmutableList;
+import org.spongepowered.api.data.type.GoldenApple;
+import org.spongepowered.api.data.type.GoldenApples;
+import org.spongepowered.common.item.SpongeGoldenApple;
+import org.spongepowered.common.registry.CatalogRegistryModule;
+import org.spongepowered.common.registry.RegisterCatalog;
+import org.spongepowered.common.registry.Registration;
+import org.spongepowered.common.registry.RegistryHelper;
+
+import java.util.Collection;
+import java.util.HashMap;
+import java.util.Map;
+import java.util.Optional;
+
+@Registration(Registration.Phase.PRE_INIT)
+@RegisterCatalog(GoldenApple.class)
+public class GoldenAppleRegistryModule implements CatalogRegistryModule<GoldenApple> {
+
+    private final Map<String, GoldenApple> goldenAppleMap = new HashMap<>();
+
+    @Override
+    public Optional<GoldenApple> getById(String id) {
+        return Optional.ofNullable(this.goldenAppleMap.get(checkNotNull(id).toLowerCase()));
+    }
+
+    @Override
+    public Collection<GoldenApple> getAll() {
+        return ImmutableList.copyOf(this.goldenAppleMap.values());
+    }
+
+    @Override
+    public void registerDefaults() {
+        this.goldenAppleMap.put("golden_apple", new SpongeGoldenApple(0, "GOLDEN_APPLE"));
+        this.goldenAppleMap.put("enchanted_golden_apple", new SpongeGoldenApple(1, "ENCHANTED_GOLDEN_APPLE"));
+
+        RegistryHelper.mapFields(GoldenApples.class, this.goldenAppleMap);
+    }
+}
